+++ lib/model/user_subscription_model.dart
@@ -176,7 +176,7 @@ class Subscription {
 
 class Order {
     int? id;
-    int? amount;
+    double? amount;
     DateTime? createdAt;
     int? userId;
     String? userName;
@@ -201,7 +201,7 @@ class Order {
 
     factory Order.fromJson(Map<String, dynamic> json) => Order(
         id: json["id"],
-        amount: json["amount"],
+        amount: json["amount"].toDouble(),
         createdAt: json["created_at"] == null ? null : DateTime.parse(json["created_at"]),
         userId: json["user_id"],
         userName: json["user_name"],
